Nicer payment term label
